fix(opensanctions-ownership-graph): handle real-world data issues
- BIGINT for IMO/MMSI columns (some MMSI values exceed INT range)
- Extract ALL entity schemas (not just 5 primary types) so relationship
  FKs always resolve — incomplete data on a PI club is better than no data
- Savepoint-based FK violation handling for relationships referencing
  entities that only exist within relationship schemas

Verified against full 2.4GB dataset: 3M entities, 120K relationships,
10.5K vessel links loaded successfully.

+++ db/migrations/025_opensanctions_graph.sql
@@ -49,8 +49,8 @@ CREATE UNIQUE INDEX IF NOT EXISTS idx_os_relationships_unique
 -- =============================================================
 CREATE TABLE IF NOT EXISTS os_vessel_links (
     entity_id    TEXT NOT NULL REFERENCES os_entities(entity_id) ON DELETE CASCADE,
-    imo          INTEGER,
-    mmsi         INTEGER,
+    imo          BIGINT,
+    mmsi         BIGINT,
     confidence   REAL NOT NULL DEFAULT 1.0,
     match_method TEXT NOT NULL,   -- 'imo_exact', 'mmsi_exact', 'name_exact'
     PRIMARY KEY (entity_id, match_method)
